- Move interceptors to app/core/interceptors - Use new mechanism to inject services
