+++ src/app/auth/auth.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { RecaptchaModule } from 'ng-recaptcha';
 import { RecaptchaFormsModule } from 'ng-recaptcha/forms';
-import { HttpCustomClientModule } from 'app/http/http-custom-client.module';
 import { SharedModule } from 'app/shared/shared.module';
 import { AuthService, UserSession, UserData } from 'app/auth/shared';
 import { SignInComponent } from './signin/signin.component';
@@ -18,7 +17,6 @@ import { Equals2 } from './password-reset/equals2.directive';
         RouterModule,
         RecaptchaModule,
         RecaptchaFormsModule,
-        HttpCustomClientModule,
         SharedModule
     ],
     providers: [
