+++ src/app/file/file.module.ts
@@ -1,5 +1,4 @@
 import { AgGridModule } from 'ag-grid-angular';
-import { HttpCustomClientModule } from 'app/http/http-custom-client.module';
 import { MarkdownModule } from 'ngx-markdown';
 import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
@@ -24,7 +23,6 @@ import { AccordionModule } from 'ngx-bootstrap';
 @NgModule({
     imports: [
         MarkdownModule.forChild(),
-        HttpCustomClientModule,
         RouterModule,
         SharedModule,
         AgGridModule.withComponents([
