+++ src/app/http/http-custom-client.module.ts
@@ -1,8 +1,5 @@
 import { NgModule } from '@angular/core';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
-import { RequestStatusService, RequestStatusServiceFactory } from './request-status.service';
-import { AuthInterceptorService } from './auth-interceptor.service';
-// import { TestBackendInterceptor } from './test-backend-interceptor.service';
+import { HttpClientModule } from '@angular/common/http';
 
 @NgModule({
     imports: [
@@ -11,27 +8,7 @@ import { AuthInterceptorService } from './auth-interceptor.service';
     exports: [
         HttpClientModule
     ],
-    providers: [
-        {
-            provide: RequestStatusService,
-            useFactory: RequestStatusServiceFactory
-        },
-        {
-            provide: HTTP_INTERCEPTORS,
-            useExisting: RequestStatusService,
-            multi: true
-        },
-        {
-            provide: HTTP_INTERCEPTORS,
-            useClass: AuthInterceptorService,
-            multi: true
-        }/*,
-        {
-            provide: HTTP_INTERCEPTORS,
-            useClass: TestBackendInterceptor,
-            multi: true
-        }*/
-    ]
+    providers: []
 })
 export class HttpCustomClientModule {
 }
